Use SDF in the HDDA
This introduces a bug if applied to `N3`, it seems that vdb_read is
non-deterministic which is very strange.

The bug has some randomness in it. Best guess right now is either
something in the decompression of the `.vdb` file that gives
unpredictable results. Either way, it seems the SDF is not correct for
`N3`.

Issue #13